feat(*): 判题服务完善

+++ zkoj-common/src/main/java/indi/snowmeow/zkoj/base/common/enums/SolutionStatusEnum.java
@@ -17,13 +17,13 @@ public enum SolutionStatusEnum {
     MLE(9),
     PE(10);
 
-    private int id;
+    private long id;
 
-    SolutionStatusEnum(int id) {
+    SolutionStatusEnum(long id) {
         this.id = id;
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
 }
